Adapt to OnlyIn removal
NeoForge 21.7.3 has removed OnlyIn from production. This fixes the crash on startup when Architectury API is used on this version.
