+++ gradle.properties
@@ -19,7 +19,7 @@ fabric_api_version=0.128.1+1.21.7
 mod_menu_version=15.0.0-beta.1
 
 forge_version=51.0.0
-neoforge_version=21.7.1-beta
+neoforge_version=21.7.3-beta
 
 # Set to empty if not snapshots
 neoforge_pr=
